feat: push host one level up for zod validation at runtime when connecting
