+++ src/icrc-wallet.spec.ts
@@ -10,7 +10,7 @@ import {
   mockLocalRelyingPartyPrincipal
 } from './mocks/call-utils.mocks';
 import {mockLocalIcRootKey} from './mocks/custom-http-agent-responses.mocks';
-import {RelyingPartyWalletOptions} from './types/relying-party-wallet-options';
+import {RelyingPartyOptions} from './types/relying-party-options';
 import {JSON_RPC_VERSION_2} from './types/rpc';
 import * as callUtils from './utils/call.utils';
 
@@ -40,7 +40,7 @@ vi.mock('@dfinity/agent', async (importOriginal) => {
 });
 
 describe('icrc-wallet', () => {
-  const mockParameters: RelyingPartyWalletOptions = {
+  const mockParameters: RelyingPartyOptions = {
     url: 'https://test.com',
     host: 'http://localhost:8080'
   };
